requirements_simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
langgraph==0.0.20
langchain==0.1.0
